move the code out of processors and into defaults
