+++ src/sentry/projects/project_rules/creator.py
@@ -9,8 +9,8 @@ from rest_framework.request import Request
 from sentry.models.project import Project
 from sentry.models.rule import Rule, RuleSource
 from sentry.types.actor import Actor
+from sentry.workflow_engine.defaults.detectors import ensure_default_detectors
 from sentry.workflow_engine.migration_helpers.issue_alert_migration import IssueAlertMigrator
-from sentry.workflow_engine.processors.detector import ensure_default_detectors
 from sentry.workflow_engine.utils.legacy_metric_tracking import report_used_legacy_models
 
 logger = logging.getLogger(__name__)
